requirements.txt
# Core Python Libraries
requests>=2.27.1
pandas>=1.4.2
pyyaml>=6.0

# API and Network
aiohttp>=3.8.1      # For asynchronous API calls
python-dotenv>=0.20.0  # For environment variable management

# Data Processing
numpy>=1.22.3

# Logging and Debugging
logging>=0.5.1.2

# Optional: For advanced data manipulation
pyarrow>=8.0.0      # Faster pandas operations

# Development and Testing
pytest>=7.1.2
pylint>=2.13.5
mypy>=0.950

# Type Checking
types-requests>=2.27.16
types-PyYAML>=6.0.7


